Show faves, comments, views and public state on the group listing page
Show faves, comments, views and public state on the group listing page

+++ FlickrPublish/PhotoCollectionController.m
@@ -92,7 +92,7 @@ static NSString * const reuseIdentifier = @"PhotoCell";
     // Configure the cell
     Photo* p = [self.photos objectAtIndex:indexPath.item];
     cell.imageTitle.text = p.name;
-    cell.imageViews.text = [NSString stringWithFormat:@"%ld", p.views];
+    cell.imageViews.text = [NSString stringWithFormat:@"%ld", (long)p.views];
     cell.thumbnailSmall.image = [UIImage imageWithData:p.imageData];
     
     return cell;
